Do static_cast for double to int64_t
Some compiler flags complain about this.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4904582
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90242}

+++ include/v8-platform.h
@@ -1232,7 +1232,7 @@ class Platform {
    * required.
    */
   virtual int64_t CurrentClockTimeMilliseconds() {
-    return floor(CurrentClockTimeMillis());
+    return static_cast<int64_t>(floor(CurrentClockTimeMillis()));
   }
 
   /**
